+++ libs/java/dal/db/dal-db/src/main/java/ca/bc/gov/tno/dal/db/entities/Claim.java
@@ -96,7 +96,7 @@ public class Claim extends AuditColumns {
    * @param name    Unique name
    * @param version Row version value
    */
-  public Claim(int id, String name, int version) {
+  public Claim(int id, String name, long version) {
     this(id, name);
     this.setVersion(version);
   }
