+++ libs/java/dal/db/dal-db/src/main/java/ca/bc/gov/tno/dal/db/entities/Role.java
@@ -109,7 +109,7 @@ public class Role extends AuditColumns {
    * @param name    Unique name
    * @param version Row version value
    */
-  public Role(int id, String name, int version) {
+  public Role(int id, String name, long version) {
     this(id, name);
     this.setVersion(version);
   }
