+++ libs/java/dal/db/dal-db/src/main/java/ca/bc/gov/tno/dal/db/entities/Tag.java
@@ -99,7 +99,7 @@ public class Tag extends AuditColumns {
    * @param name    Unique name
    * @param version Row version value
    */
-  public Tag(String id, String name, int version) {
+  public Tag(String id, String name, long version) {
     this(id, name);
     this.setVersion(version);
   }
